requirements.txt
setuptools~=61.2.0
numpy~=1.21.5
pillow~=9.0.1
scikit-image~=0.19.2
torch_optimizer
termcolor
git+https://github.com/jizongFox/tqdm.git
pyyaml~=6.0
scikit-learn
matplotlib~=3.5.1
pandas~=1.4.2
tensorboardx
scipy
opencv-python
medpy~=0.4.0
decorator
distributed
requests
randaugment
packaging
nibabel
nilearn
git+https://github.com/jizongFox/easydict.git
tensorboard
loguru~=0.6.0
gdown
git+https://github.com/jizongFox/rising.git
einops~=0.4.1
segmentation-models-pytorch
six~=1.16.0
omegaconf~=2.2.2
typing_extensions~=4.1.1
prettytable
